+++ src/natcap/invest/hra.py
@@ -551,8 +551,8 @@ def execute(args):
                 f"  Missing from criteria table: {missing_from_criteria_table}"
             )
 
-    criteria_df = pandas.read_csv(composite_criteria_table_path,
-                                  index_col=False)
+    criteria_df = utils.read_csv_to_dataframe(
+        composite_criteria_table_path, index_col=False)
     # Because criteria may be spatial, we need to prepare those spatial inputs
     # as well.
     spatial_criteria_attrs = {}
@@ -1829,8 +1829,8 @@ def _parse_criteria_table(criteria_table_path, target_composite_csv_path):
     """
     # This function requires that the table is read as a numpy array, so it's
     # easiest to read the table directly.
-    table = pandas.read_csv(criteria_table_path, header=None, sep=None,
-                            engine='python').to_numpy()
+    table = utils.read_csv_to_dataframe(
+        criteria_table_path, header=None).to_numpy()
 
     # clean up any leading or trailing whitespace.
     for row_num in range(table.shape[0]):
@@ -2379,8 +2379,8 @@ def _override_datastack_archive_criteria_table_path(
         the data dir.
     """
     args_key = 'criteria_table_path'
-    criteria_table_array = pandas.read_csv(
-        criteria_table_path, header=None, sep=None, engine='python').to_numpy()
+    criteria_table_array = utils.read_csv_to_dataframe(
+        criteria_table_path, header=None).to_numpy()
     contained_data_dir = os.path.join(data_dir, f'{args_key}_data')
 
     known_rating_cols = set()
